% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_bootstrap.R
\name{cmi_fp_bootstrap}
\alias{cmi_fp_bootstrap}
\title{Multiple, parametric conditional mean imputation for a censored covariate}
\usage{
cmi_fp_bootstrap(
  imputation_model,
  dist,
  analysis_model,
  data,
  type = "original",
  B = 10
)
}
\arguments{
\item{imputation_model}{Imputation model formula (or coercible to formula), a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See the documentation for \code{Surv} for details.}

\item{dist}{Imputation model distribution passed through to \code{survreg}. See \code{survreg} documentation for more details.}

\item{analysis_model}{Analysis model formula (or coercible to formula), a formula expression as for other regression models. The response should be a continuous outcome for normal linear regression.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{type}{String name of the type of imputation procedure to be used. Options include \code{"original"} (the default), \code{"analytical"}, \code{"stabilized (with mean)"}, and \code{"stabilized (without mean)"}.}

\item{B}{numeric, number of imputations. Default is \code{10}.}
}
\value{
A dataframe containing the pooled coefficients and standard errors for \code{analysis_model}
}
\description{
Multiple, parametric conditional mean imputation for a censored covariate using parametric survival regression to estimate the conditional survival function.
}
