% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_eq15_single.R
\name{cmi_fp_eq15_single}
\alias{cmi_fp_eq15_single}
\title{Single, fully parametric imputation for a right-censored covariate with conditional means following Equation (15)}
\usage{
cmi_fp_eq15_single(
  imputation_formula,
  dist,
  W,
  Delta,
  data,
  maxiter = 100,
  use_cumulative_hazard = FALSE
)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula) passed through to \code{survreg}, a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See \code{survreg} documentation for more details.}

\item{dist}{imputation model distribution passed through to \code{survreg}. Options of \code{dist =} \code{"exponential"}, \code{"extreme"}, \code{"gaussian"}, \code{"logistic"}, \code{"loglogistic"}, \code{"loggaussian"}, \code{"lognormal"}, \code{"rayleigh"}, \code{"t"}, and \code{"weibull"} currently accepted.}

\item{W}{character, column name for observed values of the censored covariate}

\item{Delta}{character, column name for censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{data}{dataframe or named matrix containing columns \code{W}, \code{Delta}, and any other variables in \code{imputation_formula}.}

\item{use_cumulative_hazard}{logical, if \code{TRUE} the survival function is transformed to the cumulative hazard before integration. Default is \code{FALSE}.}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{100}.}
}
\value{
A list containing:
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Single, fully parametric conditional mean imputation for a right-censored covariate using an accelerated failure-time model to estimate the conditional survival function and then integrates over the estimated survival function from \code{0} to \code{W} to compute conditional means, as in Equation (15) of the manuscript.
}
