% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_eq2_interval.R
\name{cmi_fp_eq2_interval}
\alias{cmi_fp_eq2_interval}
\title{Single, fully parametric imputation for an interval-censored covariate with conditional means following Equation (2)}
\usage{
cmi_fp_eq2_interval(imputation_formula, dist, L, U, Delta, data, maxiter = 100)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula) passed through to \code{survreg}, a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See \code{survreg} documentation for more details.}

\item{dist}{imputation model distribution passed through to \code{survreg}. See \code{survreg} documentation for more details.}

\item{L}{character, column name for the lower limit of the censoring interval for the censored covariate}

\item{U}{character, column name for the upper limit of the censoring interval for the censored covariate}

\item{Delta}{character, column name for censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{100}.}
}
\value{
A list containing:
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Single, fully parametric conditional mean imputation for an interval-censored covariate using an accelerated failure-time model to estimate the conditional survival function and then integrates over the estimated survival function from \code{W} to \code{Inf} to compute conditional means, as in Equation (1) of the manuscript.
}
