% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_eq6.R
\name{cmi_fp_eq6}
\alias{cmi_fp_eq6}
\title{Single, fully parametric conditional mean imputation for a right-censored covariate (Weibull distribution) with conditional means following Equation (6)}
\usage{
cmi_fp_eq6(
  imputation_formula,
  W,
  Delta,
  data,
  infinite_integral = TRUE,
  maxiter = 100
)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula), a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See the documentation for \code{Surv} for details.}

\item{W}{character, column name for observed values of the censored covariate}

\item{Delta}{character, column name for censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{100}.}
}
\value{
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Single, fully parametric conditional mean imputation for a right-censored covariate using an accelerated failure-time model with a Weibull distribution to estimate the conditional survival function and then uses an analytic solution to compute conditional means, as in Equation (6) of the manuscript.
}
