% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_original_single.R
\name{cmi_fp_original_single}
\alias{cmi_fp_original_single}
\title{Single, fully parametric imputation for a right-censored covariate with conditional means following Equation (1)}
\usage{
cmi_fp_original_single(imputation_model, dist, data, maxiter = 100)
}
\arguments{
\item{imputation_model}{imputation model formula (or coercible to formula) passed through to \code{survreg}, a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See \code{survreg} documentation for more details.}

\item{dist}{imputation model distribution passed through to \code{survreg}. See \code{survreg} documentation for more details.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and any other variables in \code{imputation_model}.}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{100}.}
}
\value{
A list containing:
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
\item{aic}{Akaike information criterion (AIC) from the \code{imputation_model} fit.}
\item{bic}{Bayesian information criterion (BIC) from the \code{imputation_model} fit.}
\item{coefficients}{Vector of coefficients from the \code{imputation_model} fit.}
\item{scale}{Scale from the \code{imputation_model} fit.}
}
\description{
Single, fully parametric conditional mean imputation for a right-censored covariate using an accelerated failure-time model to estimate the conditional survival function and then integrates over the estimated survival function from \code{W} to \code{Inf} to compute conditional means, as in Equation (1) of the manuscript.
}
