% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_pool_fit.R
\name{cmi_fp_pool_fit}
\alias{cmi_fp_pool_fit}
\title{Fit analysis model to each imputed dataset and pool the results using Rubin's rules}
\usage{
cmi_fp_pool_fit(analysis_model, mult_imp, rubins_rules = TRUE)
}
\arguments{
\item{analysis_model}{imputation model formula (or coercible to formula) passed through to \code{survreg}, a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See \code{survreg} documentation for more details.}

\item{mult_imp}{list (or list of lists) of imputed datasets returned from one of the \code{cmi_fp} functions}

\item{rubins_rules}{logical, if \code{rubins_rules = TRUE} (the default) Rubin's Rules are used to calculate the pooled variance. If \code{rubins_rules = FALSE}, the approximation from Bernhardt et al. (2014) is used instead.}
}
\value{
A dataframe containing the pooled coefficients and standard errors for \code{analysis_model}
}
\description{
Fit analysis model to each imputed dataset and pool the results using Rubin's rules
}
