% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_analytical.R
\name{cmi_fp_analytical}
\alias{cmi_fp_analytical}
\title{Fully parametric conditional mean imputation for a right-censored covariate with conditional means using the analytical solution}
\usage{
cmi_fp_analytical(
  imputation_formula,
  dist,
  W,
  Delta,
  data,
  maxiter = 100,
  boots = 0,
  seed = 123
)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula), a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See the documentation for \code{Surv} for details.}

\item{dist}{imputation model distribution passed through to \code{survreg}. See \code{survreg} documentation for more details.}

\item{W}{character, column name for observed values of the censored covariate.}

\item{Delta}{character, column name for censoring indicators. Note that values of zero in \code{Delta} are interpreted as censored observations.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and any other variables in \code{imputation_formula}.}

\item{boots}{(optional) numeric, for multiple imputation supply the desired number of imputations (obtained via bootstrapping) to \code{boots}. Default is \code{0}, which is single imputation.}

\item{seed}{(optional) numeric, for multiple imputation set the random seed for the bootstrapping with \code{seed}. Default is \code{123}.}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{max_iter = 100}.}
}
\value{
A list (or, in the case of multiple imputation, a list of lists) containing:
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Fully parametric conditional mean imputation for a right-censored covariate using an accelerated failure-time model to estimate the conditional survival function and then uses an analytical solution to compute conditional means.
}
