% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_eq1.R
\name{cmi_fp_eq1}
\alias{cmi_fp_eq1}
\title{Fully parametric imputation for a right-censored covariate with conditional means following Equation (1)}
\usage{
cmi_fp_eq1(
  imputation_formula,
  dist,
  W,
  Delta,
  data,
  maxiter = 100,
  boots = 0,
  seed = 123
)
}
\arguments{
\item{imputation_formula}{imputation model formula (or coercible to formula) passed through to \code{survreg}, a formula expression as for other regression models. The response is usually a survival object as returned by the \code{Surv} function. See \code{survreg} documentation for more details.}

\item{dist}{imputation model distribution passed through to \code{survreg}. See \code{survreg} documentation for more details.}

\item{W}{character, column name for observed values of the censored covariate}

\item{Delta}{character, column name for censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and any other variables in \code{imputation_formula}.}

\item{boots}{(optional) numeric, for multiple imputation supply the desired number of imputations (obtained via bootstrapping) to \code{boots}. Default is \code{0}, which is single imputation.}

\item{seed}{(optional) numeric, for multiple imputation set the random seed for the bootstrapping with \code{seed}. Default is \code{123}.}

\item{max_iter}{(optional) numeric, maximum iterations allowed in call to \code{survival::survreg()}. Default is \code{100}.}
}
\value{
A list (or, in the case of multiple imputation, a list of lists) containing:
\item{imputed_data}{A copy of \code{data} with added column \code{imp} containing the imputed values.}
\item{code}{Indicator of algorithm status (\code{TRUE} or \code{FALSE}).}
}
\description{
Fully parametric conditional mean imputation for a right-censored covariate using an accelerated failure-time model to estimate the conditional survival function and then integrates over the estimated survival function from \code{W} to \code{Inf} to compute conditional means, as in Equation (1) of the manuscript.
}
