% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp_pool_fit.R
\name{cmi_fp_pool_fit}
\alias{cmi_fp_pool_fit}
\title{Fit and pool models fit to multiply imputed datasets}
\usage{
cmi_fp_pool_fit(formula, family = gaussian, imp_data)
}
\arguments{
\item{formula}{analysis model formula (or coercible to formula), a formula expression as for other regression models.}

\item{family}{(optional) analysis model family, to be supplied to \code{glm()} to fit the model. Default is \code{family = gaussian}, but see \code{?family} for more options and details.}

\item{imp_data}{A list of lists returned from one of the following functions: \code{cmi_fp_original()}, \code{cmi_fp_stabilized()}, or \code{cmi_fp_analytical()}.}
}
\value{
A dataframe of coefficient and standard error estimates
}
\description{
Fully parametric conditional mean imputation for a right-censored covariate using an accelerated failure-time model to estimate the conditional survival function and then uses an analytical solution to compute conditional means.
}
